docs: move test files to same folder with the component
