+++ react-ts-app/src/components/Home/Home.test.tsx
@@ -1,5 +1,5 @@
 import { render, screen } from '@testing-library/react';
-import Home from '../components/Home/Home';
+import Home from './Home';
 
 describe('Home', () => {
   it('Renders heading', () => {
